Fix the burn time (deprecated API) and remove 17 dangling Javadocs
1. CZAS PALENIA BYL LICZONY PRZESTARZALYM API (realny blad).
   burnTicksOf() pytal statyczna mape AbstractFurnaceBlockEntity.getFuel()
   (oznaczona jako @Deprecated), ktora zna WYLACZNIE paliwa wpisane per-item.
   Gdy wpisu nie bylo, nasz fallback zwracal rowne 200 tickow - a wiec paliwo
   zdefiniowane tagiem albo dodane danymi przez inny mod spalalo sie
   wielokrotnie szybciej, niz powinno (wegiel to 1600 tickow).
   Teraz idziemy TA SAMA sciezka co waniliowy piec:
   ItemStack.getBurnTime(RecipeType.SMELTING) - ona przechodzi przez wartosci
   paliw z danych (FuelValues) i rozwija tagi. Sprawdzone kompilacja, ze ta
   metoda jest dostepna na naszym classpath; oba wywolania sa po stronie
   serwera, wiec nie ma problemu z danymi po stronie klienta.
   Przy okazji zniklo ostatnie uzycie przestarzalego API w projekcie.

2. ZERO ZACZEPIONYCH JAVADOKOW (bylo 17).
   Kompilacja z -Xlint:all pokazala 17 dokumentow, ktore nie sa przypisane do
   ZADNEJ deklaracji. Powstaly przy refaktorach: metoda zniknela albo
   przeniosla sie, a jej dokument zostal i opisywal nieistniejace zachowanie.
   Najgrozniejszy byl w VelocePipeNetworkManager: opisywal "Tablice polaczen
   miedzy sieciami" - mechanizm, ktory usunalem jako martwy w tym samym
   tygodniu. Ktoś czytajacy kod mialby prawo wierzyc, ze sieci moga sie laczyc.
   Inne opisywaly usuniete tlo cache craftowalnosci, stary sposob rysowania
   liczb "+N", stara wersje insertu zwracajaca boolean i indeks receptur
   trzymany w IdentityHashMap.

   Nie usunalem wszystkiego na slepo - czesc dokumentow miala prawdziwych
   wlascicieli i zostala PRZENIESIONA, bo ich metody nie mialy zadnego opisu:
     * "Wklada ile sie da i zwraca RESZTE" -> insertItemLeftover()
       (w ConnectedEndpointInfo i RefinedStorageHelper - ten opis tlumaczy
       historyczny blad duplikacji, wiec byl zbyt cenny, zeby go skasowac),
     * opis partii itemow -> countCraftableBatch(),
     * "Czy zostawic hotbar gracza dzialajacy" -> keepPlayerHotbar(),
     * "Zastepuje sloty gracza..." -> suppressPlayerSlots(),
     * dwa opisy snapshotStock scalone w jeden (drugi zaczynal sie od
       "Jak wyzej" i bez pierwszego nie mial sensu).

3. Drobiazgi: usuniety nieuzywany import, zbedny cast (int), linie z samymi
   spacjami.

Weryfikacja: javac -Xlint:all daje teraz ZERO ostrzezen z naszych plikow poza
8x [this-escape]. Kazde z tych osmiu sprawdzilem po kolei - to standardowy
wzorzec Minecrafta (registerDefaultState i stateDefinition.any() w konstruktorze
bloku, addSlot() w menu, addListener() na kontenerze block entity), a nie blad.
Przy okazji diff potwierdza, ze w zadnym pliku nie zniknela linia KODU -
usuwane byly wylacznie komentarze.

+++ src/com/craftingveloce/client/gui/VeloceCraftingTableScreen.java
@@ -1,6 +1,5 @@
 package com.craftingveloce.client.gui;
 
-
 import com.craftingveloce.network.CraftingTableToggleItemPKT;
 import com.craftingveloce.network.CraftingTableCycleRecipePKT;
 import com.mojang.blaze3d.systems.RenderSystem;
@@ -52,7 +51,6 @@ public class VeloceCraftingTableScreen extends VeloceCreativeScreen {
     private Set<Item> disabledItems;
     private Map<Item, ResourceLocation> preferredRecipes;
 
-    /** Cache: item -> lista receptur (id + wynik + skladniki). */
     /**
      * Indeks receptur po stronie klienta.
      *
@@ -130,7 +128,6 @@ public class VeloceCraftingTableScreen extends VeloceCreativeScreen {
         List<ClientRecipe> recipes = getCraftableItems().get(stack.getItem());
         return recipes != null && !recipes.isEmpty();
     }
-
 
     /**
      * Buduje indeks itemow craftowalnych z receptur dostepnych po stronie klienta.
@@ -286,7 +283,6 @@ public class VeloceCraftingTableScreen extends VeloceCreativeScreen {
     //   i modyfikowac go bezposrednio. Wtedy zakladka ekwipunku moze pokazac
     //   bufor craftera bez walki z vanilla - mamy pelna kontrole nad ukladem.
     //   Nie robic tego przez podmiane slotow ani przez drugi ekran.
-
 
     // ------------------------------------------------------------------
     // Tooltip
@@ -475,7 +471,9 @@ public class VeloceCraftingTableScreen extends VeloceCreativeScreen {
         }
         ResourceLocation current = preferredRecipes.get(item);
         int idx = current == null ? -1 : ids.indexOf(current);
-        int next = (int) (((idx + 1) % ids.size() + ids.size()) % ids.size());
+        // Podwojny modulo, zeby wynik byl nieujemny takze dla idx = -1
+        // (item bez wybranej receptury). Wyrazenie jest juz typu int.
+        int next = ((idx + 1) % ids.size() + ids.size()) % ids.size();
         ResourceLocation chosen = ids.get(next);
         preferredRecipes.put(item, chosen);
         PacketDistributor.sendToServer(
